requirements.txt
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
defusedxml==0.6.0
idna==2.10
networkx==2.5
numpy==1.18.5
onnx==1.8.0
onnxruntime==1.6.0
Pillow==8.0.1
protobuf==3.14.0
requests==2.25.1
six==1.15.0
test-generator==0.1.1
typing-extensions==3.7.4.3
urllib3==1.26.2
torch==1.7.1
opencv-python==4.4.0.46
torch==1.7.1
torchvision==0.8.2